+++ src/giit/build.py
@@ -15,18 +15,16 @@ class Build(object):
     def __init__(self, step,
                  repository,
                  build_path=None,
-                 clean_build=False,
                  data_path=None,
                  json_config=None,
-                 source_branch=None):
+                 remote_branch=None):
 
         self.step = step
         self.repository = repository
         self.build_path = build_path
-        self.clean_build = clean_build
         self.data_path = data_path
         self.json_config = json_config
-        self.source_branch = source_branch
+        self.remote_branch = remote_branch
 
     def run(self):
 
@@ -70,7 +68,7 @@ class Build(object):
         log.debug('build_path=%s', self.build_path)
         log.debug('data_path=%s', self.data_path)
         log.debug('json_config=%s', self.json_config)
-        log.debug('source_branch=%s', self.source_branch)
+        log.debug('remote_branch=%s', self.remote_branch)
 
         log.info('Lets go: %s', self.step)
 
@@ -85,9 +83,14 @@ class Build(object):
             self.build_path = os.path.join(giit_path, 'build',
                                            git_repository.unique_name)
 
-        if self.clean_build:
+        if self.step == 'clean':
+
+            log.info("Cleaning: %s", self.build_path)
+
             if os.path.isdir(self.build_path):
                 shutil.rmtree(self.build_path, ignore_errors=True)
+
+            return
 
         if not os.path.isdir(self.build_path):
             os.makedirs(self.build_path)
@@ -147,7 +150,7 @@ class Build(object):
     def resolve_factory(self):
         return giit.factory.resolve_factory(
             data_path=self.data_path,
-            source_branch=self.source_branch)
+            remote_branch=self.remote_branch)
 
     def clone_factory(self, unique_name):
         return giit.factory.cache_factory(