+++ test/test_variables_reader.py
@@ -17,7 +17,7 @@ def test_variables():
 
     context = {
         'scope': 'tag',
-        'name': '1.0.0',
+        'checkout': '1.0.0',
         'build_path': '/tmp/build',
         'source_path': '/tmp/clone'
     }
@@ -40,7 +40,7 @@ def test_variables_not_found():
 
     context = {
         'scope': 'tag',
-        'name': '1.0.0',
+        'checkout': '1.0.0',
         'build_path': '/tmp/build',
         'source_path': '/tmp/clone'
     }
@@ -58,7 +58,7 @@ def test_variables_empty():
 
     context = {
         'scope': 'tag',
-        'name': '1.0.0',
+        'checkout': '1.0.0',
         'build_path': '/tmp/build',
         'source_path': '/tmp/clone'
     }